First look at a big pile of hacks to support GDK
Caveats:
- You need to install the game from the store before GDK builds can be installed by the launcher. This applies separately for release vs preview as they're two different games.
- Add an exclusion for C:\XboxGames to Windows Defender if you don't want installing to take 1000 years
- Importing preview .msixvc is not currently supported

God knows what else currently may not work

Thank you to the following people who were super helpful in getting this done:
- @rtm516 for his work on the XVC download system, which saves a ton of
  hassle downloading the MSIXVCs
- @conmaster2112, @TwistedAsylumMC and @SmokeyStack for answering
  questions and helping me figure stuff out
- @LukeFZ for his work reverse engineering XVCs

Without all of these people this would've taken significantly longer and much more frustration.

Due to the way GDK works, it's simpler to just install the game into the XboxGames directory
and then copy the files out. This is a bit janky, but it seems like the best way for now. If
someone can figure out how to get the CIKs from MS in a reliable way, that would allow us to
extract the package directly without installing it, but for now this will have to do.

You may see some random PowerShell windows and GameInput.msi popups when installing via the
launcher. While this might look dodgy it's currently unavoidable because the only easy way
to copy out the un-encrypted .exe is using Invoke-CommandInDesktopPackage, which again is very janky.