Bidirectional data migration This only works as long as: - only a single user has worlds in the GDK data folder - Minecraft is not already running
Minecraft's own UWP data migration is leveraged for UWP -> GDK by
deleting uwpMigration.dat, which triggers importing UWP data into the
GDK locations. We do this because we don't know which of the numbered
folders to put user worlds in, since we don't know the user's current
account from the launcher's scope. We may be able to solve this in the
future, but for now this is good enough.

options.txt seems not to be preserved properly when switching versions,
which is yet to be investigated.